Revert "fix(sfu): apply screen bitrate hints to VP9"
This reverts commit 22c164cd73001808a11e7aa23af44952fdc7098e.
